Update spiders to generate json from docs pages
