+++ docs-crawler/docs/docs/items.py
@@ -12,3 +12,10 @@ class DocsItem(scrapy.Item):
     # define the fields for your item here like:
     # name = scrapy.Field()
     pass
+
+
+class ApiItem(scrapy.Item):
+    code = scrapy.Field()
+    function_name = scrapy.Field()
+    args = scrapy.Field()
+    kwargs = scrapy.Field()
