+++ docs-crawler/docs/docs/spiders/tfjs_spider.py
@@ -1,9 +1,12 @@
+import re
 import scrapy
+from docs.items import ApiItem
 from w3lib.html import remove_tags
 
 
 class TfjsSpider(scrapy.Spider):
     name = "tfjs"
+    split_def = re.compile('^([\w\.]+)\((.*)\)$')
 
     def start_requests(self):
         urls = [
@@ -23,8 +26,23 @@ class TfjsSpider(scrapy.Spider):
                 .replace('function', '')\
                 .replace('method', '')
             defs.append(text)
-        
-        with open('tfjs-functions.txt', 'w') as f:
-            f.write('\n'.join(defs))
+
+        for text in defs:
+            split = self.split_def.match(text)
+            if split is None:
+                return
+            
+            function_name = split.groups()[0].split('.')[-1]
+            params = split.groups()[1].split(',')
+            args = [p for p in params if '=' not in p]
+            kwargs = [p.split('=') for p in params if '=' in p]
+
+            item = ApiItem()
+            item['code'] = text
+            item['function_name'] = function_name
+            item['args'] = args
+            item['kwargs'] = kwargs
+            yield item
+
 
         print('\n'.join(defs))
