+++ docs-crawler/docs/docs/spiders/torch_spider.py
@@ -1,9 +1,12 @@
+import re
 import scrapy
+from docs.items import ApiItem
 from w3lib.html import remove_tags
 
 
 class TorchSpider(scrapy.Spider):
     name = "torch"
+    split_def = re.compile('^([\w\.]+)\(([\w\,\s=\*\.]*)\)')
 
     def start_requests(self):
         urls = [
@@ -18,8 +21,22 @@ class TorchSpider(scrapy.Spider):
         for selector in fdef:
             text = remove_tags(selector.get()).replace('\n', '')
             defs.append(text)
-        
-        with open('torch-functions.txt', 'w') as f:
-            f.write('\n'.join(defs))
+
+        for text in defs:
+            split = self.split_def.match(text)
+            if split is None:
+                return
+            
+            function_name = split.groups()[0].split('.')[-1]
+            params = split.groups()[1].split(',')
+            args = [p for p in params if '=' not in p]
+            kwargs = [p.split('=') for p in params if '=' in p]
+
+            item = ApiItem()
+            item['code'] = text
+            item['function_name'] = function_name
+            item['args'] = args
+            item['kwargs'] = kwargs
+            yield item
 
         print('\n'.join(defs))
